Make the library fork safe
We record the pid when establishing the connection, and compare
the recorded pid with the current one whenever we re-use the
connection.

If we detect a change, we clone the connection and close the
old one.

We make sure not to send `QUIT` when closing the old connection as
it would break the parent's connection.

We also use `dup2` for the same reason so that we can close the connection
without sending FIN or RST.